Refactor RangeAnalyzer
